% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat.R, R/utils.R
\name{asDistObject}
\alias{asDistObject}
\alias{as_dist_object}
\title{Converts a distance vector into an object of class \code{dist}}
\usage{
asDistObject(x, n, method)

as_dist_object(x, n, method)
}
\arguments{
\item{x}{data vector}

\item{n}{length of x}

\item{method}{method description}
}
\description{
Converts a distance vector into an object of class \code{dist}
}
