% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asDistObject}
\alias{asDistObject}
\title{Converts a distance vector into an object of class \code{dist}}
\usage{
asDistObject(x, n, method)
}
\description{
Converts a distance vector into an object of class \code{dist}
}
