% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casebasedreasoning.R
\docType{package}
\name{CaseBasedReasoning-package}
\alias{CaseBasedReasoning}
\alias{CaseBasedReasoning-package}
\title{CaseBasedReasoning: Case Based Reasoning}
\description{
Case-based reasoning is a problem-solving methodology that involves solving a new problem by referring to the solution of a similar problem in a large set of previously solved problems. The key aspect of Case Based Reasoning is to determine the problem that "most closely" matches the new problem at hand. This is achieved by defining a family of distance functions and using these distance functions as parameters for local averaging regression estimates of the final result. The optimal distance function is chosen based on a specific error measure used in regression estimation. This approach allows for efficient problem-solving by leveraging past experiences and adapting solutions from similar cases. The underlying concept is inspired by the work of Dippon J. et al. (2002) \doi{10.1016/S0167-9473(02)00058-0}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sipemu/case-based-reasoning}
  \item Report bugs at \url{https://github.com/sipemu/case-based-reasoning/issues}
}

}
\author{
\strong{Maintainer}: Simon Mueller \email{simon.mueller@muon-stat.com}

Other contributors:
\itemize{
  \item PD Dr. Juergen Dippon \email{juergen.dippon@mathematik.uni-stuttgart.de} [contributor]
}

}
\keyword{internal}
