<<<<<<< HEAD:man/cbr.Rd
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\docType{package}
\name{cbr}
\alias{cbr}
\alias{cbr-package}
\title{case based reasoning}
\description{
case based reasoning
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\docType{package}
\name{CaseBasedReasoning}
\alias{CaseBasedReasoning}
\alias{CaseBasedReasoning-package}
\title{case based reasoning}
\description{
case based reasoning
}
>>>>>>> 1d4213d69cd9ee46e409be573717b96a75494b2e:man/CaseBasedReasoning.Rd
