% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cbrCoxModel-Class.R
\docType{class}
\name{cbrCoxModel}
\alias{cbrCoxModel}
\title{Case Based Reasoning with Cox Model for distance calculation}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrCoxModel
}
\description{
Case Based Reasoning with Cox Model for distance calculation
}
\keyword{Cox}
\keyword{Model}

