% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrData-Class.R
\docType{data}
\name{cbrData}
\alias{cbrData}
\title{transforms data to integer representation;
necessary for c++ functions}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cbrData
}
\description{
transforms data to integer representation;
necessary for c++ functions
}
\keyword{datasets}
