% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrDistance-Class.R
\docType{class}
\name{cbrDistance}
\alias{cbrDistance}
\title{R6 reference class}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrDistance
}
\description{
R6 reference class
}
\keyword{Cox}
\keyword{Model}

