% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModels-Class.R
\docType{class}
\name{cbrRegressionModel}
\alias{cbrRegressionModel}
\title{Regression Model}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrRegressionModel$new
}
\arguments{
\item{learning}{data set for learning the Cox model}

\item{queryData}{(optional): Query data set. For each case in the query data, 
we are looking for the k (=1,…,l) similar cases in the learning data. 
Learning and query datasets need the same structure (variable names and scales)}

\item{learnVars}{(Default: all variables except endPoint): A character vector 
variable names. This variables are used for learning the model. Do not 
include time2event and event variable here. 
fitting.}

\item{endPoint}{(Default: c("Time2Event", "Event")): A character vector with
the name of the dependent variable}

\item{impute}{(Default: FALSE): Missing value imputation. Actually, not 
implemented for the regression model.}
}
\description{
Regression beta coefficients are use for building a weighted distance measure between 
the learning and verum data set. The learning data set is used for learning the
model (OLS, LR, and Cox) and use the obtained weights for calculating a (n x m)-distance 
matrix, where n is the number of cases in the learning data set and m is the 
number of cases of the query data. This distance matrix can then be used for 
cluster analysis or for getting for each case in the query data k (=1,...,l)
smilar cases from the learning data. The rms-package is used for model fitting,
variable selection, and checking the assumptions. 
If query data is ommitted, a n x n- distance matrix is returned.
}
\section{Fields}{

\describe{
\item{\code{new}}{: Initialization of the cbrRegressionModel}

\item{\code{variable_selection}}{: performs a fast backward variable selection}

\item{\code{learn}}{: Fit regression model on the learning data set}

\item{\code{check_ph}}{: Check proportional hazard assumption}

\item{\code{calc_distance_matrix}}{: Calculates full n x m / (n x n)-distance matrix}

\item{\code{calc_similar_cases}}{: get for each case in verum data nCases (=1,...,l < n) 
similar cases from the learning data; (1:nCases matching)}
}}
\keyword{Beta}

