% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cbrRFProxy-Class.R
\docType{class}
\name{cbrRFProxy}
\alias{cbrRFProxy}
\title{R6 class}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrRFProxy
}
\description{
R6 class
}
\keyword{Cox}
\keyword{Model}

