% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_Matrix_Order}
\alias{fast_Matrix_Order}
\title{Order distance matrix}
\usage{
fast_Matrix_Order(mDist, nCases)
}
\arguments{
\item{mDist}{distance matrix}

\item{nCases}{number of cases}
}
\value{
matrix with nearest cases
}
\description{
Order distance matrix
}

