% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_Vector_Order}
\alias{fast_Vector_Order}
\title{Order distance vector}
\usage{
fast_Vector_Order(vX, n)
}
\arguments{
\item{mDist}{distance matrix}

\item{nCases}{number of cases}
}
\value{
matrix with nearest cases
}
\description{
Order distance vector
}

