% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_Distance_Matrix}
\alias{get_Distance_Matrix}
\title{calculate weighted distance matrix}
\usage{
get_Distance_Matrix(newData, refData, weights)
}
\arguments{
\item{newData}{new data}

\item{refData}{reference data}
}
\value{
weights weights
}
\description{
calculate weighted distance matrix
}

