% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_nearest_Elements}
\alias{get_nearest_Elements}
\title{calculate weighted distance matrix and get nearest elements}
\usage{
get_nearest_Elements(mNew, mRef, vWeights, nCases)
}
\arguments{
\item{newData}{new data}

\item{refData}{reference data}
}
\value{
weights weights
}
\description{
calculate weighted distance matrix and get nearest elements
}

