% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrManifoldLearning-Class.R
\docType{class}
\name{manifoldLearning}
\alias{manifoldLearning}
\title{R6 reference class for getting the similar cases}
\format{An \code{\link{R6Class}} generator object}
\usage{
manifoldLearning
}
\description{
R6 reference class for getting the similar cases
}
\keyword{Cox}
\keyword{Model}

