% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.RFModel}
\alias{predict.RFModel}
\title{Predict method for RFModel}
\usage{
\method{predict}{RFModel}(object, newdata, k = 1, ...)
}
\arguments{
\item{object}{An \code{RFModel} R6 object}

\item{newdata}{Query data of class data.frame}

\item{k}{Number of similar cases to return}

\item{...}{Additional arguments (currently unused)}
}
\value{
A data.frame of similar cases
}
\description{
Predict method for RFModel
}
