% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.CoxModel}
\alias{print.CoxModel}
\title{Print method for CoxModel}
\usage{
\method{print}{CoxModel}(x, ...)
}
\arguments{
\item{x}{A \code{CoxModel} R6 object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for CoxModel
}
