% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.LinearModel}
\alias{print.LinearModel}
\title{Print method for LinearModel}
\usage{
\method{print}{LinearModel}(x, ...)
}
\arguments{
\item{x}{A \code{LinearModel} R6 object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for LinearModel
}
