% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.LogisticModel}
\alias{print.LogisticModel}
\title{Print method for LogisticModel}
\usage{
\method{print}{LogisticModel}(x, ...)
}
\arguments{
\item{x}{A \code{LogisticModel} R6 object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for LogisticModel
}
