% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrRanger.R
\name{rangerProximityMatrix}
\alias{rangerProximityMatrix}
\title{Get proximity matrix of an ranger object}
\usage{
rangerProximityMatrix(rf, x)
}
\arguments{
\item{rf}{\code{ranger} object}

\item{x}{a new dataset}
}
\value{
a \code{dist} object with pairwise proximity of observations in x
}
\description{
Get proximity matrix of an ranger object
}
\examples{
\dontrun{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
rangerProximityMatrix(rf, iris[, -5])
}

}

