% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionModel.R
\docType{data}
\name{RegressionModel}
\alias{RegressionModel}
\title{Root class for Regression Models, e.g., CPH, logistic, and linear regression}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
RegressionModel
}
\description{
Root class for Regression Models, e.g., CPH, logistic, and linear regression
}
\keyword{data-preparation}
