% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cbrSimCases-Class.R
\docType{class}
\name{simCases}
\alias{simCases}
\title{Case Based Reasoning Get Similar Cases}
\format{An \code{\link{R6Class}} generator object}
\usage{
simCases
}
\description{
Case Based Reasoning Get Similar Cases
}
\keyword{Cox}
\keyword{Model}

