% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.CoxModel}
\alias{summary.CoxModel}
\title{Summary method for CoxModel}
\usage{
\method{summary}{CoxModel}(object, ...)
}
\arguments{
\item{object}{A \code{CoxModel} R6 object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for CoxModel
}
