% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.LinearModel}
\alias{summary.LinearModel}
\title{Summary method for LinearModel}
\usage{
\method{summary}{LinearModel}(object, ...)
}
\arguments{
\item{object}{A \code{LinearModel} R6 object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for LinearModel
}
