% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.LogisticModel}
\alias{summary.LogisticModel}
\title{Summary method for LogisticModel}
\usage{
\method{summary}{LogisticModel}(object, ...)
}
\arguments{
\item{object}{A \code{LogisticModel} R6 object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for LogisticModel
}
