% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.RFModel}
\alias{summary.RFModel}
\title{Summary method for RFModel}
\usage{
\method{summary}{RFModel}(object, ...)
}
\arguments{
\item{object}{An \code{RFModel} R6 object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for RFModel
}
