% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrRanger.R
\name{terminalNodeIdsRanger}
\alias{terminalNodeIdsRanger}
\title{Get terminal node IDs of observations of an ranger object}
\usage{
terminalNodeIdsRanger(rf, x)
}
\arguments{
\item{rf}{\code{ranger} object}

\item{x}{a new dataset}
}
\value{
Matrix with terminal node IDs for all observations in x (rows) and
        trees (columns)
}
\description{
Get terminal node IDs of observations of an ranger object
}
\examples{
\dontrun{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
terminalNodeIdsRanger(rf, iris[, -5])
}

}

