<<<<<<< HEAD:man/cbrValidate.Rd
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrValidate-Class.R
\docType{class}
\name{cbrValidate}
\alias{cbrValidate}
\title{R6 Validation Class for case based reasoning}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrValidate
}
\description{
R6 Validation Class for case based reasoning
}
\keyword{validation}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\docType{class}
\name{Validate}
\alias{Validate}
\title{R6 Validation Class for case based reasoning}
\format{An \code{\link{R6Class}} generator object}
\usage{
Validate
}
\description{
R6 Validation Class for case based reasoning
}
\keyword{validation}
>>>>>>> 1d4213d69cd9ee46e409be573717b96a75494b2e:man/Validate.Rd
