% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cbrDistance-Class.R
\docType{class}
\name{cbrDistance}
\alias{cbrDistance}
\title{Case Based Reasoning distance calculation}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrDistance
}
\description{
Case Based Reasoning distance calculation
}
\keyword{Cox}
\keyword{Model}

