% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxBetaModel.R
\name{CoxBetaModel}
\alias{CoxBetaModel}
\title{Cox-Beta Model for Case-Based-Reasoning}
\description{
Cox-Beta Model for Case-Based-Reasoning

Cox-Beta Model for Case-Based-Reasoning
}
\details{
Regression beta coefficients obtained from a CPH regression model fitted on the 
training data are used for building a weighted distance measure between
train and test data. Afterwards, we will use these weights for calculating a 
(n x m)-distance matrix, where n is the number of observations in the training data, 
and m is the number of observations of the test data. The user can use this 
distance matrix for further cluster analysis or for extracting for each test observation 
k (= 1,...,l) similar cases from the train data. We use the rms-package for model fitting,
variable selection, and checking model assumptions.
If the user omits the test data, this functions returns a n x n-distance matrix.
}
\section{Super class}{
\code{\link[CaseBasedReasoning:CBRBase]{CaseBasedReasoning::CBRBase}} -> \code{CoxBetaModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{the statistical model}

\item{\code{model_params}}{rms arguments}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-check_ph}{\code{CoxBetaModel$check_ph()}}
\item \href{#method-clone}{\code{CoxBetaModel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="calc_distance_matrix">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-calc_distance_matrix}{\code{CaseBasedReasoning::CBRBase$calc_distance_matrix()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="fit">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-fit}{\code{CaseBasedReasoning::CBRBase$fit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="get_similar_cases">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-get_similar_cases}{\code{CaseBasedReasoning::CBRBase$get_similar_cases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="initialize">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-initialize}{\code{CaseBasedReasoning::CBRBase$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_ph"></a>}}
\if{latex}{\out{\hypertarget{method-check_ph}{}}}
\subsection{Method \code{check_ph()}}{
Check proportional hazard assumption graphically
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxBetaModel$check_ph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoxBetaModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
