% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticBetaModel.R
\name{LogisticBetaModel}
\alias{LogisticBetaModel}
\title{Logistic Regression Model for Case-Based-Reasoning}
\description{
Logistic Regression Model for Case-Based-Reasoning

Logistic Regression Model for Case-Based-Reasoning
}
\section{Super classes}{
\code{\link[CaseBasedReasoning:CBRBase]{CaseBasedReasoning::CBRBase}} -> \code{\link[CaseBasedReasoning:RegressionModel]{CaseBasedReasoning::RegressionModel}} -> \code{LogisticBetaModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{the statistical model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{LogisticBetaModel$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="calc_distance_matrix">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-calc_distance_matrix}{\code{CaseBasedReasoning::CBRBase$calc_distance_matrix()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="get_similar_cases">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-get_similar_cases}{\code{CaseBasedReasoning::CBRBase$get_similar_cases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="initialize">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-initialize}{\code{CaseBasedReasoning::CBRBase$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="RegressionModel" data-id="fit">}\href{../../CaseBasedReasoning/html/RegressionModel.html#method-fit}{\code{CaseBasedReasoning::RegressionModel$fit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="RegressionModel" data-id="print">}\href{../../CaseBasedReasoning/html/RegressionModel.html#method-print}{\code{CaseBasedReasoning::RegressionModel$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="RegressionModel" data-id="variable_selection">}\href{../../CaseBasedReasoning/html/RegressionModel.html#method-variable_selection}{\code{CaseBasedReasoning::RegressionModel$variable_selection()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticBetaModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
