% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfDistance.R
\name{proximityMatrixRanger}
\alias{proximityMatrixRanger}
\title{Get proximity matrix of an ranger object}
\usage{
proximityMatrixRanger(x, rf)
}
\arguments{
\item{x}{a new dataset}

\item{rf}{\code{ranger} object}
}
\value{
a \code{dist} object with pairwise proximity of observations in x
}
\description{
Get proximity matrix of an ranger object
}
\examples{
\dontrun{
require(ranger)
rf <- ranger(Species ~ ., data = iris, num.trees = 5, write.forest = TRUE)
proximityMatrixRanger(iris[, -5], rf)
}

}

