% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionModel.R
\name{RegressionModel}
\alias{RegressionModel}
\title{Root class for Regression Models, e.g., CPH, logistic, and linear regression}
\description{
Root class for Regression Models, e.g., CPH, logistic, and linear regression

Root class for Regression Models, e.g., CPH, logistic, and linear regression
}
\keyword{data-preparation}
\section{Super class}{
\code{\link[CaseBasedReasoning:CBRBase]{CaseBasedReasoning::CBRBase}} -> \code{RegressionModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model_params}}{rms arguments}

\item{\code{weights}}{Weights for distance calculation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{RegressionModel$print()}}
\item \href{#method-variable_selection}{\code{RegressionModel$variable_selection()}}
\item \href{#method-fit}{\code{RegressionModel$fit()}}
\item \href{#method-clone}{\code{RegressionModel$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="calc_distance_matrix">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-calc_distance_matrix}{\code{CaseBasedReasoning::CBRBase$calc_distance_matrix()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="get_similar_cases">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-get_similar_cases}{\code{CaseBasedReasoning::CBRBase$get_similar_cases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="CaseBasedReasoning" data-topic="CBRBase" data-id="initialize">}\href{../../CaseBasedReasoning/html/CBRBase.html#method-initialize}{\code{CaseBasedReasoning::CBRBase$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints information of the initialized object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionModel$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-variable_selection"></a>}}
\if{latex}{\out{\hypertarget{method-variable_selection}{}}}
\subsection{Method \code{variable_selection()}}{
Fast backward variable selection with penalization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionModel$variable_selection(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Training data of class data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the RandomForest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionModel$fit(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Training data of class data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
