<<<<<<< HEAD:man/cbrRFModel.Rd
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrRFModel-Class.R
\docType{class}
\name{cbrRFModel}
\alias{cbrRFModel}
\title{RandomForest Proximity}
\format{An \code{\link{R6Class}} generator object}
\usage{
cbrRFModel
}
\arguments{
\item{formula}{: formula for learning the Cox model}

\item{data}{: the dataset for learning the model}

\item{queryData}{(optional) : Query data set. For each case in the query data,
we are looking for the k (=1,…,l) similar cases in the learning data.
Learning and query datasets need the same structure (variable names and scales)}
}
\description{
This class uses the proximity matrix of the random survival forest algorithm 
as a similarity matrix (sqrt(1 - proximity matrix)) of learning and verum 
cases. By default all cases with at least one missing values are dropped 
from learning, calculating the distance matrix, and searching for similar
cases.
}
\keyword{Cox}
\keyword{Model}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFModel.R
\docType{class}
\name{RFModel}
\alias{RFModel}
\title{RandomForest Proximity}
\format{An \code{\link{R6Class}} generator object}
\usage{
RFModel
}
\arguments{
\item{formula}{: formula for learning the Cox model}

\item{data}{: the dataset for learning the model}

\item{queryData}{(optional) : Query data set. For each case in the query data,
we are looking for the k (=1,…,l) similar cases in the learning data.
Learning and query datasets need the same structure (variable names and scales)}
}
\description{
This class uses the proximity matrix of the random survival forest algorithm 
as a similarity matrix (sqrt(1 - proximity matrix)) of learning and verum 
cases. By default all cases with at least one missing values are dropped 
from learning, calculating the distance matrix, and searching for similar
cases.
}
\keyword{Cox}
\keyword{Model}
>>>>>>> 1d4213d69cd9ee46e409be573717b96a75494b2e:man/RFModel.Rd
